requirements.txt
numpy==1.16.2
Pillow==2.2.1